server/requirements.txt
alembic==1.14.1
asttokens==3.0.0
backcall==0.2.0
bcrypt==4.3.0
blinker==1.8.2
click==8.1.8
decorator==5.2.1
dnspython==2.6.1
email_validator==2.2.0
executing==2.2.0
Faker==35.2.2
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipdb==0.13.9
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.41
SQLAlchemy-serializer==1.4.12
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.3
typing_extensions==4.13.2
wcwidth==0.2.13
Werkzeug==2.2.2
zipp==3.20.2
gunicorn==21.2.0
python-dotenv==1.0.1
